Add event's name validation
We've been using the event's name to know what kind of event it is.

I think we should add the name presence validator,
what do you think about it?

If should be more validations, please let me know, and also, if the
validation is not needed/desired feel free to close the pr.